Rename LightningTree to LightningTreeNode
As per our code style, the file name should be the same as the class name.

Contributes to issue CURA-8550.
